Reworked Characteristic
Some of the main complaints people had were about having to reiderate through the char[] more times than neeeded. I agreed, and have made it so my program will never iterate more than once. Another issue was that there was unneeded confusion with allocating new data for the cleaned version of the string, and so I scrapped all of that and did everything inside of the characteristic function. This was done as a test, but I think it reduces confusion. Another major change was including the process of building the integer while checking the validity of the char []. I also included a possible issue where input is given but would overflow integer max if computed.